+++ src/ServerMessage.scala
@@ -4,7 +4,7 @@ case class OrderSuccess(orderId: Int) extends ServerMessage {
   override def toString = "OK " + orderId
 }
 case class OrderFailure() extends ServerMessage {override def toString = "FAIL"}
-case class Executed(orderId: Int, amount: Int, price: Double) extends ServerMessage {
+case class Executed(orderId: Int, amount: Int, price: BigDecimal) extends ServerMessage {
   override def toString = "DONE " + orderId + " " + amount + " " + price
 }
 
@@ -17,7 +17,7 @@ case class LoginFailure() extends ServerMessage {override def toString = "FAIL"}
 case class AdminSuccess() extends ServerMessage {override def toString = "OK"}
 case class AdminFailure() extends ServerMessage {override def toString = "FAIL"}
 
-case class BalanceMessage(balance: Double) extends ServerMessage {override def toString = balance toString}
+case class BalanceMessage(balance: BigDecimal) extends ServerMessage {override def toString = balance toString}
 case class ClientOrdersList(orders: Iterable[Order]) extends ServerMessage {
   override def toString = {
     (for (order <- orders) yield
@@ -27,7 +27,7 @@ case class ClientOrdersList(orders: Iterable[Order]) extends ServerMessage {
   }
 }
 
-case class LastTradedPrice(price: Option[Double]) extends ServerMessage {
+case class LastTradedPrice(price: Option[BigDecimal]) extends ServerMessage {
   override def toString = price match {
     case Some(p) => p toString
     case None    => "NONE"
